drivers: flash: stm32_qspi: Fix status register access for dual-flash
When dual-flash mode is enabled, two identical flash memories are
connected to the QUADSPI peripheral, each having its own set of
registers. This means that when reading or writing a flash register,
this has to be made for both flash memories.

For example, when reading a status register (1 byte), the QUADSPI
peripheral must be configured to read two bytes of data, which
correspond respectively to the value of the register in the first and
second flash memory. Same thing when writing.

Before this commit, when dual-flash mode was enabled, only the register
of the first flash memory was considered, which means the second flash
memory could be incorrectly configured and that any write/erase
operation could be considered as completed too early, if the operation
takes more time to complete for the second flash memory.

(cherry picked from commit c4afaaacf6c19096a0f106d2da217132fa8ac90d)

Original-Signed-off-by: Thomas Altenbach <altenbach.thomas@gmail.com>
GitOrigin-RevId: c4afaaacf6c19096a0f106d2da217132fa8ac90d
Cr-Build-Id: 8708556658338805137
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708556658338805137
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I07e143d119c5b3420c8d2c85d55665f4b737422b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6781302
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>