drivers: eth_nxp_enet_qos: Move RX underrun code
Extract the RX underrun handle code to a separate function.

Named the function like "dma_rx_resume" to make it clear what the
function is really supposed to do.

Demote the error about not being a first descriptor to a warning.
Because most likely we already got an error about something else which
caused us to drop the packet in the first place. The rest of the frames
are expected to be dropped. And make the string shorter.

Also remove the debug message because the control bits do not tell us
any more information than we don't know already. They only tell us that
we own the descriptor (known since we are processing the frame), that it
is not a first descriptor (known since that is the reason we would drop
it at this point, as indicated by the warning).

(cherry picked from commit 6267ec59effaf443671c251ebb4fb0313f49ad99)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 6267ec59effaf443671c251ebb4fb0313f49ad99
Cr-Build-Id: 8707288298771325409
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8707288298771325409
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6adae7254c7f553a4781a5fab154764f04c6c570
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6823715
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>