trusted-firmware-m: Set --align when signing
The current version of TF-M script that sign MCUboot image uses a
default alignment of 1. This value varies between flash devices
and not all accept the default 1. This improve the script picking
the write-block-size property from the current flash controller
and pass as the --align parameter when signing an image.

devices in the Zephyr tree and an exception will throw when
a device is not supported.

(cherry picked from commit debb59830b7e8eb7f8a6470b29bbdcc9f291695a)

Original-Note: This solution works out-of-box for the vast majority of
Original-Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>
GitOrigin-RevId: debb59830b7e8eb7f8a6470b29bbdcc9f291695a
Cr-Build-Id: 8700108490016058097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700108490016058097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I87c27d28db5e2d76ee800f5477b0f47e311dbfea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7083529
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>