trusted-firmware-m: Set --confirm when signing
The current behavior when signing an image adds --pad but does not
confirm the image. This appears to be a mistake, as the user should
inspect the image status in the Firmware Upgrade software. If an image
is not --confirmed, the FSM cannot infer the correct states. This sets
the image as confirmed to resolve the issue.

(cherry picked from commit 6cee10c5b92639d18e35a7ff27e47a460c25cfc0)

Original-Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>
GitOrigin-RevId: 6cee10c5b92639d18e35a7ff27e47a460c25cfc0
Cr-Build-Id: 8700108490016058097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700108490016058097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ifaf1d6e7468a76ea4eadea3c830676d5c7753397
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7083532
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>