samples: net: sockets: http_server: Dont block system workq
Since the netstats handler calls on functions which reach deep into the
networking stack, there is a lot of points actually at which it can be
blocked, even forever. So having this handler on the system workqueue is
not a good idea and can even cause a deadlock in some cases if it's
blocked waiting on a synchronization primitive that would be given by a
work item scheduled later in the queue. Therefore, make a workqueue
specifically for this http server socket instead of using the system
one.

(cherry picked from commit fb0ca1653f0ccd6923f723cd565490568223316c)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: fb0ca1653f0ccd6923f723cd565490568223316c
Cr-Build-Id: 8708352815478829489
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708352815478829489
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8502ba44cd6f35be1867fa2c3ad45dc04e4a1617
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6788936
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>