Bluetooth: Classic: SDP: Improve UUID 128 support
In current implementation, the UUID 128 is not well supported. The
found issue includes,
Issue 1: No clear byte order of the UUID 128 in local SDP record.
Issue 2: No clear byte order of the UUID 128 of the SDP discover
parameter.

For issue 1,
Add the description to note that if the SDP attribute type is
`BT_SDP_UINT128`, `BT_SDP_INT128`, and `BT_SDP_UUID128`, the byte
order should be little-endian.

And swap the 128bit from little-endian to big-endian when responding
the peer SDP discovery request.

For issue 2,
Add the description to note that if the SDP discovery type is
`Service Search` and `Service Search Attribute`, and UUID is UUID 128,
the passed UUID data of the discovery request parameter should be
represented as the little-endian byte-order sequence.

And swap the 128bit from little-endian to big-endian when packing the
SDP discovery packet.

(cherry picked from commit af42036ca6692fb58fdbefac5ce79a4bb6edcfe9)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: af42036ca6692fb58fdbefac5ce79a4bb6edcfe9
Cr-Build-Id: 8708013075837114577
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708013075837114577
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4ef4608420e5660b4ac339f6c95454d9c8a0f770
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6797741
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>