drivers: hwspinlock: sqn: use cluster id instead of core mpidr
Using the mpidr of the core to lock the hwspinlock meant it acted as
a spinlock in the cluster itself, which is exactly the opposite of what we
tried to achieve.
Additionally, this tied the driver to the arm architecture, even though the
HW was not.

(cherry picked from commit 2c86e65e710f881e5bd550856fbc24cd9625a3f6)

Original-Signed-off-by: Pierrick Guillaume <pguillaume@fymyte.com>
GitOrigin-RevId: 2c86e65e710f881e5bd550856fbc24cd9625a3f6
Cr-Build-Id: 8701988378913007633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701988378913007633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ieed0b26218ae1873bab95639ea5ef30d4515ca2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7009404
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>