drivers: nrf_wifi: Fix defaults to avoid false recovery
In case of a busy environment and if STA is far, then we see many
retries for the frames that cause the RPU to be awake though host has
de-asserted wakeup_now signal, this leads to WDT interrupt and host
thinks that it has given sleep opportunity to RPU and initiates a
recovery.

To handle this, increase the sleep opportunity window to 5s to cover all
the retries, this solves the false recovery problem. While at it, also
increase the range, no reason to limit to lower window. And update the
help text with the warning about power consumption.

(cherry picked from commit 40b7b87ad7a754713afa05b9bb279767d7953eaa)

Original-Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>
GitOrigin-RevId: 40b7b87ad7a754713afa05b9bb279767d7953eaa
Cr-Build-Id: 8708058374682561809
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708058374682561809
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I58a0149621d596383226eecf9025f524c54d37f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6795923
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>