modules: cmsis, cmsis_6: only add the intended cmsis module
The current code base is meant to use cmsis for Cortex A and R and
cmsis_6 for Cortex M, but the build system is configured to include the
path for both when Cortex M is selected. This leaves us exposed to PR
using the old headers, that would not get caught in CI but would fail
the build on a project using Cortex M that only has the cmsis_6 module.

Change the cmsis module setting to only include the module files in the
intended case.

(cherry picked from commit 56a446b4ee04b3e8cb0a85f60e49a077a8801593)

Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: 56a446b4ee04b3e8cb0a85f60e49a077a8801593
Cr-Build-Id: 8707378897374018081
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8707378897374018081
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I819c959f06cf775f406d7d17f0062c985b1efccb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6821404
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>