Bluetooth: CTS: Fix issue with writing without subscription
The write_ct function calls bt_cts_send_notification and if
bt_cts_send_notification fails to send a notification (e.g.
if there are no clients subscribed), it would return an
error, and thus send a ATT error response to the client,
even if the value was actually written.

This commit changes the behavior so that the notification
does not affect the return value (as it is not directly
linked to the write itself), but rather just logs it if
it fails to send the notification.

The notification can failed to be send for 2 main reasons:
1) The application returns an error on fill_current_cts_time
2) bt_gatt_notify fails to notify due to missing buffers

The first we cannot do much about. For the latter we can
either just live with the missing notification, or implement
a retry if we run out of buffers. For this simple fix the
former was chosen, but can be revisited if it turns out
to be a significant problem.

(cherry picked from commit eef2aa88b32f48a16d722759009d8c56a104942f)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: eef2aa88b32f48a16d722759009d8c56a104942f
Cr-Build-Id: 8707288298771325409
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8707288298771325409
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I520afa81742e5305f2be4d35fcffec185d380978
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6823718
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>