kernel: message does not execute correct put front behavior
When the buffer is full, Thread A gets pended (blocked).
If Thread B later calls the get function, it will unpend Thread A,
allowing it to resume and put the message into the queue.
In this situation, we need to know whether Thread A should
continue with put to front or put to end.

In order to resolve this issue, we don't allow set timeout
parameter for `k_msgq_put_front` and this parameter is always
`K_NO_WAIT`.

(cherry picked from commit d361ec9692856d90333d6761a862885564cc15a9)

Original-Signed-off-by: TaiJu Wu <tjwu1217@gmail.com>
GitOrigin-RevId: d361ec9692856d90333d6761a862885564cc15a9
Cr-Build-Id: 8703460579189758961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703460579189758961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3396c553adc4023ef152caa52a61d8a2dea982ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6952469
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>