adc: Remove CONFIG_ADC_MCUX_ETC
This Kconfig does not belong in the ADC folder, because there is not
actually a zephyr ADC driver for this. Also, remove HAS_MCUX_ADC_ETC as
well because it is a useless config.

The cmake line to pull in this driver from the SDK in the zephyr repo is
totally unnecessary. If a user wants to use this SDK driver they can add
it to their build like any other SDK driver or any other
external code module. Zephyr should not be a cesspool of random build
glue for random pieces of code.

(cherry picked from commit 5e3a432fedaa6b00cd7a18e27d35e154c487a4ba)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 5e3a432fedaa6b00cd7a18e27d35e154c487a4ba
Cr-Build-Id: 8706201137051610721
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8706201137051610721
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4bd4fe58616dbfc8efd7b0b3ac88f76fc4f1421a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6858092
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>