samples: ensure struct addrinfo is always zero initialized
There's few samples with a struct addrinfo on the stack that only
initialize some of the fields, feeding whatever is on the stack for the
rest to the getaddrinfo() code. Bumped into an issue where that
sometimes happens to set the bit of the (ai_flags & AI_NUMERICHOST),
making host resolution failing.

Add a couple of memset on the affected samples so this does not happen.

(cherry picked from commit d9f2f70133336f617f4376e071406cf3cc8abe06)

Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: d9f2f70133336f617f4376e071406cf3cc8abe06
Cr-Build-Id: 8707288298771325409
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8707288298771325409
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1461304ab14174d6df75b1c2c8025af0bdee1882
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6823737
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>