Bluetooth: BAP: Sink: Remove LOG_WRN for started/stopped
Removed the LOG_WRN if the started or stopped callback
for streams are not set. The reasoning is that there is also
the group-wide started and stopped callbacks, and an application
may want to use those instead of the per-stream callbacks.

(cherry picked from commit cb7965a1a923181727fb6071ada7bafc131ae560)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: cb7965a1a923181727fb6071ada7bafc131ae560
Cr-Build-Id: 8706540875242396545
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8706540875242396545
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I63f01c62c707bb7bf6cb2123f553aede12eb1f59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6851675
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>