intel_adsp: kconfig: Increase the default size of MEMORY_WIN_2_SIZE
The current default of 8192 will provide only 2 pages in debug window:

However, the coredump is hardwired to use slot1, which by default is not
valid.

Increase the default window size to 12288 to allow three pages.

This change affects CAVS25 only as it is using default window sizes and the
window 3 is not used in this configuration at all (it was used with IPC3
only), so we do have enough space for the three page - we could even
increase the default to cover 4 pages (8192+8192), but let's be
conservative on this.

(cherry picked from commit c7dd9d79bdf37eb30933ec2251b5c8ebb1ef7796)

Original-page0: descriptors
Original-page1: slot0
Original-Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
GitOrigin-RevId: c7dd9d79bdf37eb30933ec2251b5c8ebb1ef7796
Cr-Build-Id: 8695691889394186705
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695691889394186705
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib5b47e8e1959212fc32b043c930c913e1894a52a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7253967
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>