Samples: Fix ID for VCNL4040 sensor sample
The ID of the sample is wrong, so it is not possible to link
to the sample documentation. I found this issue when working
on a shield for the VCNL4040 proximity sensor.

To my knowledge, there are no references to the invalid ID yet.

(cherry picked from commit 8843fd9fe1800164fbb5221d95d1eba7d0699458)

Original-Signed-off-by: Jonas Berg <jonas.s.t.berg@gmail.com>
GitOrigin-RevId: 8843fd9fe1800164fbb5221d95d1eba7d0699458
Cr-Build-Id: 8706676771030746673
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8706676771030746673
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0e84b03c5122c65d7c7021c0c3756c6cba23ca41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6844922
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>