tests/arm_mpu_pxn: Insist that trivial functions are not inlined
GCC 14 takes a look these functions which are just 'return true' and
inlines them even though they has the noinline attribute set. This happens
because the compiler computes the possible range of values from the
function call, and as that is a single value (true), it replaces the call
with that value. So it's not strictly inlining the function?

Insist a bit harder by adding a compiler_barrier to the functions which
seems to solve the issue.

(cherry picked from commit b56e1385217a9497b4988e0b30eeb3407dac738f)

Original-Signed-off-by: Keith Packard <keithp@keithp.com>
GitOrigin-RevId: b56e1385217a9497b4988e0b30eeb3407dac738f
Cr-Build-Id: 8706155838053660929
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8706155838053660929
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ifd44045da2aa2930e1c34f5256b46452b6227b7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6860649
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>