display: add PIXEL_FORMAT_AL_88
When rendering an overlay on top of something like a video stream on a
monochrome display, the only way to achieve this right now is by using
ARGB8888 and discarding two of the colors, which is not ideal.

This introduces the new AL_88 pixel format, which is much more efficient
for this usecase.

This change is somewhat inspired by LVGL, which also supports AL_88
natively.

(cherry picked from commit 8c6c761744ca10e0e23bfb962893336127677a07)

Original-Signed-off-by: Martin Stumpf <finomnis@gmail.com>
GitOrigin-RevId: 8c6c761744ca10e0e23bfb962893336127677a07
Cr-Build-Id: 8706495577247152017
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8706495577247152017
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I65c4b4c3ff1e678921bfcffe1e20f0764e2f8a4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6853378
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>