drivers: mbox: imx_mu: ignore TriggerInterrupt failures
The ipc backends, like icmsg.c, don't expect errors when
using mailbox send for notifications. In some places,
icmsg.c ignores the return value of the mailbox send,
in some places it propagates the error up, and in some
places it asserts.

icmsg only uses the mailbox send with NULL msg as a means
to interrupt the other processor, and all data is passed
via shared memory. So it is safe to ignore the case when
the interrupt is already pending, since the other
processor should handle the pending message when it gets
around the handling the interrupt. As long as the order
of clearing the interrupt is before the handling of the
data, there should be no case where not pending a new
interrupt when one is already pending would result in missing
any new message.

(cherry picked from commit 04778ff6ccbbb007fe76b07df099dd073a6eed9e)

Original-Signed-off-by: Mike J. Chen <mjchen@google.com>
GitOrigin-RevId: 04778ff6ccbbb007fe76b07df099dd073a6eed9e
Cr-Build-Id: 8701558043672538497
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701558043672538497
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4cd582ca5994b438be2d92a5847709f27fd4932d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7013772
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>