Bluetooth: BAP: Fix bad use of bt_ascs_ase_status
The bt_ascs_ase_status was stored directly in the
struct bt_bap_ep, which was an issue as
bt_ascs_ase_status is a variable sized struct.
The struct could have been moved to the end of
bt_bap_ep to avoid the issues it caused, but
there is no actual reason to use that struct
rather than storing the ID and state as
seperate values.

The ascs_ep_get_state exists but was not widely
used in ascs. Modify places to use the function
instead of directly accessing the state.

(cherry picked from commit ce9e8f8b9e020b5626509c84880749bf610560e5)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: ce9e8f8b9e020b5626509c84880749bf610560e5
Cr-Build-Id: 8704208004692697937
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704208004692697937
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7595e83301faa8e03ba1aab46b45c33b53f1be98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6929267
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>