net: dns: fix the bounds conditions on the failures and error returns
The loop condition is not the count of the actual tries since the
continue or breaks can occur on the loop and thus a failure count
can falsely match a loop count even when a query is outstanding.

This can be seen when a dual stack resolve is used for IPv4 and IPv6
that has two fds to be used and the interface has only gotten an
IPv6 address. The failures on the IPv4 will indicate a 1 and the break
on the tried attempt on the IPv6 will break the loop on a 1.

(cherry picked from commit 0fe091bc4691bb89e09388446f1b1a0c5c74e1a1)

Original-Signed-off-by: Charles Hardin <ckhardin@gmail.com>
GitOrigin-RevId: 0fe091bc4691bb89e09388446f1b1a0c5c74e1a1
Cr-Build-Id: 8705951996224277857
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8705951996224277857
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I25422e55b4e65b52c196a24ad7d2e95624dce690
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6869553
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>