Bluetooth: SDP: Fix ssa discovery issue caused by setting range
If the range of the service search attribute request is set and the
start of the range is not 0x0000, the response data of the request is
not valid. It is caused by the data element sequence has not be added
to response data. The client cannot parse the response data properly.

The root cause of this issue is that the total length of service
record will be added only when adding the first attribute of service
record.

Fix the issue by reset the `state.last_att` and `state.last_att_index`
if the index of current record is not same with `state.current_svc`.

And do not update `state.last_att`, `state.last_att_index`, and
`state.current_svc` in the function `select_attrs()`.

(cherry picked from commit de0b53578a27d8cfbd45442cc73b263049a6c30c)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: de0b53578a27d8cfbd45442cc73b263049a6c30c
Cr-Build-Id: 8705838749847595073
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8705838749847595073
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I34a55892b744009f6bc7056be786604bf99197d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6875461
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>