drivers: clock_control: nrf: Fix missing clock started callback
804e502484 fix got reverted as it introduced a bug where callback
with HFCLK started indication was not called because state is not
correctly indicated. BT-dedicated API does not set state and does
not expect to get the callback when HFCLK is started. 804e502484
was intended to fix the bug that was visible on nrf54lx. It was not
visible on nrf52 because of a errata workaround which was calling
HFCLK started callback only when state was set to STARTING and that
was not the case when BT-dedicated API was used. Applying this
`workaround` approach to all cases.

(cherry picked from commit e1f9de7842cc195aef28da563df583840c52ba78)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: e1f9de7842cc195aef28da563df583840c52ba78
Cr-Build-Id: 8702758451915321377
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702758451915321377
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic679b821b68d24f778925735ce436d5fe0504454
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6983880
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>