modem: cmux: Handle C/R bit from address field
The C/R bit in the address field should be handled as explained
in 5.2.1.2 in the spec (3GPP TS 127.010).

To detect if the frame is a command or a response,
we need to know who was the initiator of the CMUX channel.

>    Initiator is the station that take the initiative to initialize
>    the multiplexer (i.e. sends the SABM command at DLCI 0 )

See the table from given section of the specification.

Also, on UIH frames 5.4.3.1 says
>    The frames sent by the initiating station have the C/R bit set to 1
>    and those sent by the responding station have the C/R bit set to 0.

(cherry picked from commit b1942ced1a9a94b52c0cd241943972b058d2c5d0)

Original-NOTE: This is different than a C/R bit in the Type field.
Original-Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>
GitOrigin-RevId: b1942ced1a9a94b52c0cd241943972b058d2c5d0
Cr-Build-Id: 8701014460813377745
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701014460813377745
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I81bad990925c2b5e38a16baa14dbac53d2afa675
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7040818
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>