soc: nxp: imx95: set and restore wakeup irq mask in pm context
The CMC interface controls the entry and exit of the
CPU's low-power mode and the identification of the wake-up source.
To ensure the normal operation of the system's low-power timing sequence,
when transfer IDLE_RUN to IDLE_SLEEP, it is necessary to ensure
that the system is not awakened by the wake-up source during this stage.
Therefore, an IRQ MASK needs to be set on the CMC

Before the CPU enters the low power mode, a wake up mask needs to be set
according to the situation where the interrupt controller is enabled
at that time.
After the cpu exits the low power mode, resume needs to be performed

(cherry picked from commit cfb1ae4efedf987e52964464b513efa76fd73dea)

Original-Signed-off-by: Yongxu Wang <yongxu.wang@nxp.com>
GitOrigin-RevId: cfb1ae4efedf987e52964464b513efa76fd73dea
Cr-Build-Id: 8704208004692697937
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704208004692697937
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Idd0586bb76cd1140cb25628ffab1ce01c8e0b3f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6929275
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>