drivers: spi: litex: litespi: flush pending irq before enable
flush pending irq before enableing the irq.

In the litex soc, if an irq is not enabled, it will
it can still be registered and will be pending
if it occurs, it will just not be send to the cpu
until the irq is enabled.

(cherry picked from commit 2e3a57fc964fbdbe752e5357d181cdfa92ab4a04)

Original-Signed-off-by: Fin Maaß <f.maass@vogl-electronic.com>
GitOrigin-RevId: 2e3a57fc964fbdbe752e5357d181cdfa92ab4a04
Cr-Build-Id: 8704117406820186449
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704117406820186449
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I30012cdb2a8d14a77d82e6c78fa44d17af58e656
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6936018
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>