Bluetooth: BAP: SD: Remove address lookups
Remove lookups in the Scan Delegator that relates to
the advertiser address. The reason for this, is that the
address is not considered a unique value for receive states,
since the address may change over time in the case of (N)RPAs.
Instead we shall rely exclusively on the address type, the sid
and the broadcast ID.

The implementation of the Scan Delegator and Broadcast Sink
has been updated to not use addresses for lookups anymore,
and there has been a minor API modification to set the
PA sync state as part of bt_bap_scan_delegator_add_src
as the higher layers are better suited to handle the
PA Sync state.

(cherry picked from commit 64ea5334e491d7ee5db5c0d78fec051b91bab9d2)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 64ea5334e491d7ee5db5c0d78fec051b91bab9d2
Cr-Build-Id: 8703460579189758961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703460579189758961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iefbecd4d5f267bd4d073c3022ee067d55345cd63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6958155
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>