boards: silabs: Remove nonfunctional wakeup trigger
Remove wakeup trigger from Series 2 boards. The implementation
seems to configure a GPIO to be driven low on boot, which has
nothing to do with wakeup. In order to actually configure a
wakeup GPIO, the GPIO_INT_WAKEUP flag would need to be used when
registering an interrupt on the pin. However, this should not
be the responsibility of the board. It is an application
concern, so remove it from the boards.

(cherry picked from commit 38e9d2cd470e1d83b535aabb7d974197b4056fa8)

Original-Signed-off-by: Aksel Skauge Mellbye <aksel.mellbye@silabs.com>
GitOrigin-RevId: 38e9d2cd470e1d83b535aabb7d974197b4056fa8
Cr-Build-Id: 8703913562775995409
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703913562775995409
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I716b156ad75983c365b2e27c087202d129d32c44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6944932
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>