boards: atmel: Fix devicetree style
The Zephyr is implementing a linter to catch devicetree problems. This
boards comments are exceptions and are formatted to be compliance with
the linter. However due to the exception the linter check was disabled
to avoid undesirable formatting and keep the original details in the
comments.

(cherry picked from commit 1e22848f6d11ad1a6e39c8505feb76d79194e460)

Original-Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>
GitOrigin-RevId: 1e22848f6d11ad1a6e39c8505feb76d79194e460
Cr-Build-Id: 8704208004692697937
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704208004692697937
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0671b05890f486817cf0b4ea1f0253e20cba4c13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6930528
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>