Bluetooth: PACS: Improve handling of notify on reconnect
When we reconnect to a bonded device we always attempted to
send the changed PACS notifications, regardless of
whether the remote device has subscribed to them or not.

This resulted in a lot of failed attempts to send
notifications when
CONFIG_BT_GATT_ENFORCE_SUBSCRIPTION is enabled (which it is
by default).

Ideally we would not set the flags for devices that have not
subscribed, but since we cannot call bt_gatt_is_subscribed for
non-connected devices, we need to do this when the notify_cb
is called.

It would be possible to use BT_GATT_CCC_MANAGED instead
of BT_AUDIO_CCC to keep track of the writes for each client,
but that would be a minor optimization which would require
significant rework and additions to PACS. The current
solution is inspired by the implementation used in has.c

(cherry picked from commit e0c5c277a5de2cf136fc58cce2990839b47bd214)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: e0c5c277a5de2cf136fc58cce2990839b47bd214
Cr-Build-Id: 8703913562775995409
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703913562775995409
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia042710623a1c222077b7e63a015c6b8d7c7eeae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6945048
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>